Store user ratings

+++ src/com/laterball/server/model/RatingSubmission.kt
@@ -1,3 +1,3 @@
 package com.laterball.server.model
 
-class RatingSubmission(val fixtureId: Int, val rating: Int, val csrf: String)
+class RatingSubmission(val fixtureId: Int, val rating: Double, val csrf: String)
